Refactor and and fix URL issue where SignIn page was using `signup` endpoint.
