--- src/App.css
@@ -1,104 +1,0 @@
-* {
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-}
-
-/* styling for the rich editor (Quill) */
-.ql-editor {
-  min-height: 400px;
-  font-size: 18px;
-}
-
-button {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-  transform: translate(0px, 0px) rotate(0deg);
-  transition: 0.2s;
-  box-shadow: -4px -2px 16px 0px #ffffff, 4px 2px 16px 0px #e8f2fe;
-  border-width: 2px 2px 2px 2px;
-  border-style: solid;
-  border-color: rgba(255, 255, 255, 0.333);
-  border-radius: 40px 40px 40px 40px;
-  background-color: #e0e8ef;
-  padding: 16px 24px 16px 28px;
-  color: #7e97b8;
-  font-weight: 500;
-  font-size: 13px;
-  font-family: inherit;
-  letter-spacing: 0.4px;
-  text-transform: uppercase;
-}
-button:hover {
-  box-shadow: -2px -1px 8px 0px #ffffff, 2px 1px 8px 0px #e9f7e6;
-  background-color: #e8f2fe;
-  color: #516d91;
-}
-
-button:active {
-  box-shadow: none;
-}
-
-.rainbow-text {
-  animation: rainbow 5.5s ease-in-out;
-  animation-iteration-count: infinite;
-  font-weight: 700;
-}
-@keyframes rainbow {
-  100%,
-  0% {
-    color: rgb(127, 0, 255);
-  }
-
-  50% {
-    color: rgb(255, 0, 255);
-  }
-
-  75% {
-    color: rgb(127, 0, 255);
-  }
-  81% {
-    color: rgb(255, 0, 127);
-  }
-  100% {
-    color: rgb(242, 0, 255);
-  }
-}
-* {
-  scrollbar-width: thin;
-  scrollbar-color: #e8f2fe #ffffff;
-}
-
-/* Chrome, Edge, and Safari */
-/* *::-webkit-scrollbar {
-    width: 16px;
-}
-
-*::-webkit-scrollbar-track {
-    background: #e9f7e6;
-}
-
-*::-webkit-scrollbar-thumb {
-    border-radius: 10px;
-    border: 3px solid black;
-} */
-
-.frame1 {
-  position: fixed;
-  top: 9.5%;
-  left: 0;
-  width: 100%;
-  height: 100vh;
-  background: #e9f7e6;
-}
-
-.frame2 {
-  position: fixed;
-  top: 9.5%;
-  left: 0;
-  width: 100%;
-  height: 100vh;
-  background: azure;
-}
