+++ src/components/BoardPostItem.tsx
