+++ src/components/CommentForm.tsx
@@ -1,5 +1,5 @@
 import React, { useContext, useState } from 'react';
-import { UserContext } from './UserContext';
+import { UserContext } from '../contexts/UserContext';
 import { motion } from 'framer-motion';
 
 const CommentForm = ({
