+++ src/pages/Post.tsx
@@ -2,12 +2,12 @@ import { motion } from 'framer-motion';
 import { useCallback, useContext, useEffect, useState } from 'react';
 import { Link, useNavigate, useParams } from 'react-router-dom';
 
-import { UserContext } from './UserContext';
+import { UserContext } from '../contexts/UserContext';
 
-import CommentForm from './CommentForm';
-import CommentList from './CommentList';
+import CommentForm from '../components/CommentForm';
+import CommentList from '../components/CommentList';
 
-import { postPageAnimation } from './animations';
+import { postPageAnimation } from '../utils/animations';
 
 export type PostType = {
   id: number;
@@ -38,7 +38,7 @@ export type ResponseCommentType = {
   updatedAt: string;
 };
 
-const PostPage = () => {
+const Post = () => {
   const { id: postId } = useParams();
   const navigate = useNavigate();
   const currentUser = useContext(UserContext);
@@ -210,4 +210,4 @@ const PostPage = () => {
   );
 };
 
-export default PostPage;
+export default Post;
